Focus blur. More efficient drawing.
We can take advantage of the fact that our image is one-dimensional
to make it just 1 pixel high, and then use `drawImage` to splat it
onto the canvas.

Add a Gaussian blurring effect with distance. Really this should
simulate the effect of "focus", so that the shape in your crosshairs
is always in focus and the others out of focus, but that can be
done later.